requirements.txt
boto3==1.6.14
botocore==1.9.14
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==1.11.11
docutils==0.14
futures==3.2.0
gTTS==1.2.2
gTTS-token==1.1.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
numpy==1.14.2
Pillow==5.0.0
pyTelegramBotAPI==3.6.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
SpeechRecognition==3.8.1
urllib3==1.22
Werkzeug==0.14.1
